# api/v1alpha1/openapi.yaml
openapi: 3.0.1
info:
  title: Flight Control API
  version: v1alpha1
  description: |
    [Flight Control](https://flightctl.io) is a service for declarative management of fleets of edge devices and their workloads.
  contact:
    name: The Flight Control Team
    url: https://flightctl.io
    email: team@flightctl.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/v1/organizations/{orgID}/enrollmentrequests:
    get:
      tags:
        - enrollmentrequest
      description: List EnrollmentRequest resources.
      operationId: listEnrollmentRequests
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: continue
          in: query
          description: An optional parameter to query more results from the server. The value of the paramter must match the value of the 'continue' field in the previous list response.
          required: false
          schema:
            type: string
        - name: labelSelector
          in: query
          description: A selector to restrict the list of returned objects by their labels. Defaults to everything.
          schema:
            type: string
        - name: fieldSelector
          in: query
          description: A selector to restrict the list of returned objects by their fields, supporting operators like '=', '==', and '!=' (e.g., "key1=value1,key2!=value2").
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of results returned in the list response. The server will set the 'continue' field in the list response if more results exist. The continue value may then be specified as parameter in a subsequent query.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequestList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      tags:
        - enrollmentrequest
      description: Create an EnrollmentRequest resource.
      operationId: createEnrollmentRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: StatusConflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/organizations/{orgID}/enrollmentrequests/{name}:
    get:
      tags:
        - enrollmentrequest
      description: Get an EnrollmentRequest resource.
      operationId: getEnrollmentRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to get.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    put:
      tags:
        - enrollmentrequest
      description: Update an EnrollmentRequest resource.
      operationId: replaceEnrollmentRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: StatusConflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    patch:
      tags:
        - enrollmentrequest
      description: Patch an EnrollmentRequest resource.
      operationId: patchEnrollmentRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to patch.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      tags:
        - enrollmentrequest
      description: Delete an EnrollmentRequest resource.
      operationId: deleteEnrollmentRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to delete.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/organizations/{orgID}/enrollmentrequests/{name}/status:
    get:
      tags:
        - enrollmentrequest
      description: Get the status of an EnrollmentRequest resource.
      operationId: getEnrollmentRequestStatus
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to get.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    put:
      tags:
        - enrollmentrequest
      description: Update the status of an EnrollmentRequest resource.
      operationId: replaceEnrollmentRequestStatus
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    patch:
      tags:
        - enrollmentrequest
      description: Patch the status of an EnrollmentRequest resource.
      operationId: patchEnrollmentRequestStatus
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest resource to patch.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/organizations/{orgID}/enrollmentrequests/{name}/approval:
    put:
      tags:
        - enrollmentrequest
      description: Approve or deny an EnrollmentRequest.
      operationId: approveEnrollmentRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the EnrollmentRequest to approve or deny.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequestApproval'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentRequestApprovalStatus'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  schemas:
...
    PatchRequest:
      type: array
      items:
        type: object
        additionalProperties: false
        required:
          - op
          - path
        properties:
          path:
            description: A JSON Pointer path.
            type: string
          value:
            description: The value to add or replace.
          op:
            description: The operation to perform.
            type: string
            enum:
              - add
              - replace
              - remove
...
    DeviceStatus:
      type: object
      description: DeviceStatus represents information about the status of a device. Status may trail the actual state of a device.
      required:
        - conditions
        - systemInfo
        - applications
        - applicationsSummary
        - resources
        - integrity
        - config
        - os
        - updated
        - summary
        - lastSeen
        - lifecycle
      properties:
        conditions:
          type: array
          description: Conditions represent the observations of a the current state of a device.
          items:
            $ref: '#/components/schemas/Condition'
        systemInfo:
          $ref: "#/components/schemas/DeviceSystemInfo"
        applications:
          type: array
          description: List of device application status.
          items:
            $ref: "#/components/schemas/DeviceApplicationStatus"
        applicationsSummary:
          $ref: "#/components/schemas/DeviceApplicationsSummaryStatus"
        resources:
          $ref: "#/components/schemas/DeviceResourceStatus"
        integrity:
          $ref: "#/components/schemas/DeviceIntegrityStatus"
        config:
          $ref: "#/components/schemas/DeviceConfigStatus"
        os:
          $ref: "#/components/schemas/DeviceOsStatus"
        updated:
          $ref: "#/components/schemas/DeviceUpdatedStatus"
        summary:
          $ref: "#/components/schemas/DeviceSummaryStatus"
        lastSeen:
          type: string
          description: The last time the device was seen by the service.
          format: date-time
        lifecycle:
          $ref: '#/components/schemas/DeviceLifecycleStatus'
...
    DeviceApplicationsSummaryStatus:
      type: object
      description: A summary of the health of applications on the device.
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/ApplicationsSummaryStatusType"
        info:
          type: string
          description: Human readable information detailing the last application transition.
...
    EnrollmentRequestApproval:
      type: object
      description: EnrollmentRequestApproval contains information about the approval of a device enrollment request.
      properties:
        labels:
          type: object
          additionalProperties:
            type: string
          description: A set of labels to apply to the device.
        approved:
          type: boolean
          description: Indicates whether the request has been approved.
      required:
        - approved
    EnrollmentRequestApprovalStatus:
      description: EnrollmentRequestApprovalStatus represents information about the status of a device enrollment request approval.
      allOf:
      - $ref: "#/components/schemas/EnrollmentRequestApproval"
      - type: object
        description: EnrollmentRequestApprovalStatus contains information about the status of approval of a device enrollment request.
        properties:
          approvedBy:
            type: string
            description: The name of the approver.
          approvedAt:
            type: string
            format: date-time
            description: The time at which the request was approved.
        required:
        - approvedBy
        - approvedAt
...
    EnrollmentRequest:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/EnrollmentRequestSpec'
        status:
          $ref: '#/components/schemas/EnrollmentRequestStatus'
      required:
        - apiVersion
        - kind
        - metadata
        - spec
      description: EnrollmentRequest represents a request for approval to enroll a device.
    EnrollmentRequestList:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '#/components/schemas/ListMeta'
        items:
          type: array
          description: 'List of EnrollmentRequest.'
          items:
            $ref: '#/components/schemas/EnrollmentRequest'
      required:
        - apiVersion
        - kind
        - metadata
        - items
      description: EnrollmentRequestList is a list of EnrollmentRequest.
    EnrollmentRequestSpec:
      required:
        - csr
      type: object
      properties:
        csr:
          type: string
          description: 'The PEM-encoded PKCS#10 certificate signing request.'
        deviceStatus:
          $ref: '#/components/schemas/DeviceStatus'
        labels:
          type: object
          additionalProperties:
            type: string
          description: 'A set of labels that the service will apply to this device when its enrollment is approved.'
      description: EnrollmentRequestSpec is a description of a EnrollmentRequest's target state.
    EnrollmentRequestStatus:
      type: object
      properties:
        certificate:
          type: string
          description: 'The PEM-encoded signed certificate.'
        conditions:
          type: array
          description: 'Current state of the EnrollmentRequest.'
          items:
            $ref: '#/components/schemas/Condition'
        approval:
          $ref: '#/components/schemas/EnrollmentRequestApprovalStatus'
      required:
        - conditions
      description: EnrollmentRequestStatus represents information about the status of a EnrollmentRequest.
...
    ListMeta:
      type: object
      properties:
        continue:
          type: string
          description: May be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message.
        remainingItemCount:
          type: integer
          description: The number of subsequent items in the list which are not included in this list response. If the list request contained label or field selectors, then the number of remaining items is unknown and the field will be left unset and omitted during serialization. If the list is complete (either because it is not chunking or because this is the last chunk), then there are no more remaining items and this field will be left unset and omitted during serialization. Servers older than v1.15 do not set this field. The intended use of the remainingItemCount is *estimating* the size of a collection. Clients should not rely on the remainingItemCount to be set or to be exact.
          format: int64
      description: ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.
    ObjectMeta:
      type: object
      properties:
        creationTimestamp:
          type: string
          description: The time the object was created.
          format: date-time
        deletionTimestamp:
          type: string
          description: The time the object will be deleted.
          format: date-time
        name:
          type: string
          description: The name of the object.
        labels:
          type: object
          description: Map of string keys and values that can be used to organize and categorize (scope and select) objects.
          additionalProperties:
            type: string
        generation:
          type: integer
          description: A sequence number representing a specific generation of the desired state. Populated by the system. Read-only.
          format: int64
        owner:
          type: string
          description: A resource that owns this resource, in "kind/name" format.
        annotations:
          type: object
          additionalProperties:
            type: string
          description: Properties set by the service.
        resourceVersion:
          type: string
          description: An opaque string that identifies the server's internal version of an object.
      description: ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.
...
    Status:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        code:
          type: integer
          format: int32
          description: Suggested HTTP return code for this status, 0 if not set.
        message:
          type: string
          description: A human-readable description of the status of this operation.
        reason:
          type: string
          description: A machine-readable description of why this operation is in the "Failure" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.
        status:
          type: string
          description: 'Status of the operation. One of: "Success" or "Failure". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.'
      description: Status is a return value for calls that don't return other objects.
      required:
        - apiVersion
        - kind
        - status
        - code
        - reason
        - message
...